requirements.txt
environs
requests
aiogram==3.4.0
sqlalchemy==2.0.25
python-dotenv
asyncpg
aiosqlite
greenlet==3.1.1
aioredis==2.0.1
sqlalchemy-utils==0.41.2
alembic
psycopg2-binary
httpx
python-docx
PyPDF2
redis
psutil